Extracted LogoutButton from household to GeneralUI | updated footer with LogoutButton
